+++ simapp/provenance.go
@@ -32,9 +32,10 @@ import (
 // It performs the following actions:
 //   - Registers the KV store keys required by the name, attribute, and marker modules.
 //   - Initializes the NameKeeper, AttributeKeeper, and MarkerKeeper using the legacy Provenance wiring pattern.
-//   - Injects the MarkerKeeper, NameKeeper, and AttributeKeeper into the VaultKeeper (via app.VaultKeeper.MarkerKeeper,
-//     app.VaultKeeper.NameKeeper, and app.VaultKeeper.AttrKeeper) to enable restricted marker management,
-//     name resolution, and attribute-based gating within the vault module.
+//   - Injects the MarkerKeeper, MetadataKeeper, NameKeeper, and AttributeKeeper into the VaultKeeper (via
+//     app.VaultKeeper.MarkerKeeper, app.VaultKeeper.MetadataKeeper, app.VaultKeeper.NameKeeper, and
+//     app.VaultKeeper.AttrKeeper) to enable restricted marker management, scope resolution, name resolution,
+//     and attribute-based gating within the vault module.
 //   - Registers the modules with the app for inclusion in BeginBlocker, EndBlocker, InitGenesis, etc.
 //
 // This function is typically called during app initialization to ensure
@@ -109,6 +110,7 @@ func (app *SimApp) RegisterProvenanceModules() error {
 	)
 
 	app.VaultKeeper.MarkerKeeper = app.MarkerKeeper
+	app.VaultKeeper.MetadataKeeper = app.MetadataKeeper
 	app.VaultKeeper.NameKeeper = app.NameKeeper
 	app.VaultKeeper.AttrKeeper = app.AttributeKeeper
 	app.VaultKeeper.ExchangeKeeper = app.ExchangeKeeper